# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Data Collection and Payment Management API",
        "description": "Production-ready API for managing users, suppliers, products, collections, rates, and payments with multi-unit tracking, versioned rates, and automated payment calculations. Built on Clean Architecture, SOLID, DRY, and KISS principles.\n *\n * **Authentication:**\n * All protected endpoints require JWT Bearer token authentication. Include the token in the Authorization header.\n *\n * **Rate Limiting:**\n * - Public endpoints (login, register): 5 requests per minute\n * - Standard protected endpoints: 60 requests per minute\n * - Report endpoints: 30 requests per minute (resource intensive)\n *\n * **Version Control:**\n * Update operations on suppliers, products, rates, collections, and payments include version conflict checking to prevent concurrent modification issues. Returns 409 on conflict.\n *\n * **Audit Logging:**\n * All authenticated requests are logged for audit purposes including user actions, IP addresses, and timestamps.",
        "contact": {
            "name": "API Support",
            "email": "support@ledger.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
...
        }
...
    "paths": {
...
        "/rates": {
            "get": {
                "tags": [
                    "Rates"
                ],
                "summary": "Get all rates",
                "description": "Retrieve versioned product rates with historical preservation",
                "operationId": "getRates",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "unit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "effective_from",
                            "enum": [
                                "rate",
                                "unit",
                                "effective_from",
                                "effective_to",
                                "version",
                                "created_at",
                                "updated_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "description": "Paginated rate list with product details",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rates"
                ],
                "summary": "Create new rate",
                "description": "Create a new versioned rate for a product with automatic version incrementing",
                "operationId": "createRate",
                "requestBody": {
                    "description": "Rate data with effective date range",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "rate",
                                    "unit",
                                    "effective_from"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "ID of the product",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rate": {
                                        "description": "Rate per unit",
                                        "type": "number",
                                        "format": "float",
                                        "example": 250
                                    },
                                    "unit": {
                                        "description": "Unit of measurement",
                                        "type": "string",
                                        "example": "kg"
                                    },
                                    "effective_from": {
                                        "description": "Rate effective from date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-01"
                                    },
                                    "effective_to": {
                                        "description": "Rate effective to date (optional)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "description": "Active status",
                                        "type": "boolean",
                                        "example": true,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rate created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rate created successfully"
                                        },
                                        "data": {
                                            "description": "Rate details with version number",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Bearer Token authentication. Include the token in the Authorization header as: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}